restructuring hid/zemu/mock
