+++ .circleci/config.yml
@@ -2,14 +2,13 @@ version: 2
 jobs:
   build:
     docker:
-      - image: golang:1.13
+      - image: golang:1.14
     environment:
       GO111MODULE: "on"
     working_directory: /ledger-go
     steps:
       - checkout
-      - run: go build -mod=readonly ledger.go apduWrapper.go
-      - run: go test
+      - run: go test -tags ledger_mock
 workflows:
   version: 2
   build_all:
