+++ Makefile
@@ -1,2 +1,2 @@
 test:
-	go test ledger_test.go ledger.go apduWrapper.go
+	go test -tags ledger_mock
