+++ apduWrapper.go
@@ -18,10 +18,47 @@ package ledger_go
 
 import (
 	"encoding/binary"
+	"fmt"
 	"github.com/pkg/errors"
 )
 
 var codec = binary.BigEndian
+
+func ErrorMessage(errorCode uint16) string {
+	switch errorCode {
+	// FIXME: Code and description don't match for 0x6982 and 0x6983 based on
+	// apdu spec: https://www.eftlab.co.uk/index.php/site-map/knowledge-base/118-apdu-response-list
+
+	case 0x6400:
+		return "[APDU_CODE_EXECUTION_ERROR] No information given (NV-Ram not changed)"
+	case 0x6700:
+		return "[APDU_CODE_WRONG_LENGTH] Wrong length"
+	case 0x6982:
+		return "[APDU_CODE_EMPTY_BUFFER] Security condition not satisfied"
+	case 0x6983:
+		return "[APDU_CODE_OUTPUT_BUFFER_TOO_SMALL] Authentication method blocked"
+	case 0x6984:
+		return "[APDU_CODE_DATA_INVALID] Referenced data reversibly blocked (invalidated)"
+	case 0x6985:
+		return "[APDU_CODE_CONDITIONS_NOT_SATISFIED] Conditions of use not satisfied"
+	case 0x6986:
+		return "[APDU_CODE_COMMAND_NOT_ALLOWED] Command not allowed (no current EF)"
+	case 0x6A80:
+		return "[APDU_CODE_BAD_KEY_HANDLE] The parameters in the data field are incorrect"
+	case 0x6B00:
+		return "[APDU_CODE_INVALID_P1P2] Wrong parameter(s) P1-P2"
+	case 0x6D00:
+		return "[APDU_CODE_INS_NOT_SUPPORTED] Instruction code not supported or invalid"
+	case 0x6E00:
+		return "[APDU_CODE_CLA_NOT_SUPPORTED] Class not supported"
+	case 0x6F00:
+		return "APDU_CODE_UNKNOWN"
+	case 0x6F01:
+		return "APDU_CODE_SIGN_VERIFY_ERROR"
+	default:
+		return fmt.Sprintf("Error code: %04x", errorCode)
+	}
+}
 
 func SerializePacket(
 	channel uint16,
